requirements.txt
# INPI Heatmap Visual MVP - Dependencies for Streamlit Cloud

# Streamlit
streamlit>=1.29.0

# Machine Learning
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0

# Image Processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Hugging Face
huggingface-hub>=0.19.0

# Configuration
python-dotenv>=1.0.0
